fix(tmaze): complete ABI migration to vecenv.h
Add required num_agents and rng fields to TMaze struct.
Change actions/terminals from int*/uchar* to float* to match
vecenv.h buffer ownership. Remove allocate_TMaze from my_init
since vecenv manages observation/action/reward/terminal buffers.
Fix dict_get call to use ->value accessor (2-arg API).
Cast float actions to int in c_step for discrete action handling.

Fixes #542 (tmaze environment only)

+++ ocean/tmaze/tmaze.c
@@ -8,12 +8,12 @@ int main() {
     c_render(&env);
     while (!WindowShouldClose()) {
         if (IsKeyDown(KEY_LEFT_SHIFT)) {
-            env.actions[0] = FORWARD;
-            if (IsKeyDown(KEY_LEFT)  || IsKeyDown(KEY_A)) env.actions[0] = LEFT;
-            if (IsKeyDown(KEY_RIGHT) || IsKeyDown(KEY_D)) env.actions[0] = RIGHT;
+            env.actions[0] = (float)FORWARD;
+            if (IsKeyDown(KEY_LEFT)  || IsKeyDown(KEY_A)) env.actions[0] = (float)LEFT;
+            if (IsKeyDown(KEY_RIGHT) || IsKeyDown(KEY_D)) env.actions[0] = (float)RIGHT;
 
         } else {
-            env.actions[0] = rand() % 3;
+            env.actions[0] = (float)(rand() % 3);
         }
         c_step(&env);
         c_render(&env);
